+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# v0.36 2017-09-02 Wekan release
+
+This release adds the following new features:
+
+* [Import attachments related activities from Wekan and
+   Trello](https://github.com/wekan/wekan/pull/1202).
+
+Thanks to GitHub user GhassenRjab for contributions.
+
 # v0.35 2017-09-02 Wekan release
 
 This release adds the following new features:
@@ -20,6 +29,7 @@ This release adds the following new features:
 Known issues:
 
 * [Attachment creation times are not imported to Trello and Wekan](https://github.com/wekan/wekan/issues/1157).
+  - This is fixed in v0.36.
 
 Thanks to GitHub user GhassenRjab for contributions.